+++ controllers/admin/page/deleteController.js
